// src/model/project.ts
/**
 * 
 * @export
 * @interface Project
 */
export interface Project {
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    version: string;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    link: string;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    tag: string;
    /**
     * 
     * @type {Date}
     * @memberof Project
     */
    timestamp: Date;
}